Refactoring (#118)
* refactored Submodule classes into separate files

* using MAE and AS functions from respective classes instead of Snakefile or R script header

* fixed input for MAE rules

* use chr with bam counts for ASEReadCounter

* added test for aberrant splicing count file functions

* FRASER compatibility fixes

* fixed subsetGroup API change
