+++ .travis.yml
@@ -22,12 +22,16 @@ install:
   - conda create -q -n drop -c conda-forge -c bioconda python=$TRAVIS_PYTHON_VERSION r-base=4.0.2
   - conda activate drop
   - conda install -c conda-forge -c bioconda drop
-  - conda remove --force drop wbuild
-  - conda install -c conda-forge -c bioconda wbuild=1.7.1
+  - conda remove --force drop
   - pip install -r requirements_test.txt
 
   # install drop
   - pip install . -vv
+
+  # install drop
+  - pip uninstall wbuild -y
+  - pip install -r requirements_test.txt
+  - pip install -e . -vv
 
 script: 
   - conda list
