+++ drop/config/__init__.py
@@ -1,3 +1,5 @@
 from .DropConfig import DropConfig
 from .SampleAnnotation import SampleAnnotation
-from .Submodules import AE, AS, MAE
+from .submodules.MonoallelicExpression import MAE
+from .submodules.AberrantSplicing import AS
+from .submodules.AberrantExpression import AE
