+++ drop/config/submodules/__init__.py
@@ -1,0 +1,4 @@
+from .Submodules import *
+from .AberrantExpression import *
+from .AberrantSplicing import *
+from .MonoallelicExpression import *
