+++ drop/modules/aberrant-splicing-pipeline/Counting/01_4_countRNA_nonSplitReads_merge.R
@@ -2,12 +2,6 @@
 #' title: Merge Nonsplit Counts
 #' author: Luise Schuller
 #' wb:
-#'  py:
-#'  - |
-#'   def getNonSplitCountFiles(dataset):
-#'       ids = sa.getIDsByGroup(dataset, assay="RNA")
-#'       file_stump = cfg.getProcessedDataDir() + f"/aberrant_splicing/datasets/cache/raw-{dataset}/sample_tmp/nonSplitCounts/"
-#'       return expand(file_stump + "sample_{sample_id}.done", sample_id=ids) 
 #'  log:
 #'    - snakemake: '`sm str(tmp_dir / "AS" / "{dataset}" / "01_4_nonSplitReadsMerge.Rds")`'
 #'  params:
@@ -15,7 +9,7 @@
 #'   - workingDir: '`sm cfg.getProcessedDataDir() + "/aberrant_splicing/datasets"`'
 #'  threads: 20
 #'  input:
-#'   - sample_counts:  '`sm lambda w: getNonSplitCountFiles(w.dataset)`'
+#'   - sample_counts:  '`sm lambda w: cfg.AS.getNonSplitCountFiles(w.dataset)`'
 #'   - gRangesNonSplitCounts: '`sm cfg.getProcessedDataDir() + 
 #'                          "/aberrant_splicing/datasets/cache/raw-{dataset}/gRanges_NonSplitCounts.rds"`'
 #'  output:
