+++ drop/modules/aberrant-splicing-pipeline/FRASER/Summary.R
@@ -62,7 +62,7 @@ topN <- 30000
 topJ <- 10000
 for(type in psiTypes){
     before <- plotCountCorHeatmap(
-        fds = fds,
+        fds,
         type = type,
         logit = TRUE,
         topN = topN,
