+++ drop/modules/mae-pipeline/MAE/ASEReadCounter.sh
@@ -8,7 +8,9 @@
 # 6 {input.fasta}
 # 7 {config[mae][gatkIgnoreHeaderCheck]}
 # 8 {output.counted}
-# 9 {config[tools][bcftoolsCmd]}
+# 9 {params.bcftools}
+#10 {params.samtools}
+#11 {params.gatk}
 
 ncbi2ucsc=$1
 ucsc2ncbi=$2
@@ -19,6 +21,8 @@ fasta=$6
 sanity=$7
 output=$8
 bcftools=$9
+samtools=${10}
+gatk=${11}
 
 tmp=$(mktemp)
 header="contig\tposition\tvariantID\trefAllele\taltAllele\t"
@@ -27,36 +31,36 @@ header+="lowBaseQDepth\trawDepth\totherBases\timproperPairs"
 echo -e $header >> $tmp
 
 # get chr format
+bam_chr=$($samtools idxstats ${bam_file} | grep -vP "\t0\t0" | cut -f1) # only chr with coverage
 vcf_chr=$($bcftools view ${vcf_file} | cut -f1 | grep -v '#' | uniq)
-if [ $(echo ${vcf_chr} | grep 'chr' | wc -l) -eq 0 ]
-then
-    echo "use NCBI format"
-    canonical=$ncbi2ucsc
+if [ "$(echo ${vcf_chr} | grep -c 'chr')" -eq 0 ]; then
+  echo "use NCBI format"
+  canonical=$ncbi2ucsc
 else
-    echo "use UCSC format"
-    canonical=$ucsc2ncbi
+  echo "use UCSC format"
+  canonical=$ucsc2ncbi
 fi
 # subset from canonical chromosomes
-chr_subset=$(comm -12  <(cut -f1 -d" " ${canonical} | sort -u) <(echo ${vcf_chr} | xargs -n1 | sort -u))
+chr_subset=$(comm -12 <(cut -f1 -d" " ${canonical} | sort -u) <(echo ${vcf_chr} | xargs -n1 | sort -u))
+chr_subset=$(comm -12 <(echo ${bam_chr} | xargs -n1) <(echo ${chr_subset} | xargs -n1) | uniq)
 
-for chr in $chr_subset
-do
-    echo $chr
-    gatk ASEReadCounter \
+for chr in $chr_subset; do
+  echo $chr
+  $gatk ASEReadCounter \
     -R ${fasta} \
     -I ${bam_file} \
     -V ${vcf_file} \
     -L ${chr} \
     --verbosity ERROR \
-    --disable-sequence-dictionary-validation ${sanity} \
-    | tail -n+2 >> $tmp
+    --QUIET true \
+    --disable-sequence-dictionary-validation ${sanity} |
+    tail -n+2 >>$tmp
 done
 
 echo $mae_id
 cat $tmp | awk -v id="${mae_id}" \
-    -F $'\t' 'BEGIN {OFS = FS} NR==1{print $0, "ID"} NR>1{print $0, id}' \
-    | bgzip > ${output}
+  -F $'\t' 'BEGIN {OFS = FS} NR==1{print $0, "ID"} NR>1{print $0, id}' |
+  bgzip >${output}
 rm ${tmp}
 
 zcat ${output} | head
-
